navigator: move loose route helpers to util/
route-bounds.ts, route-display.ts, route-features.ts were sitting
loose at the top of src/ as peers of create-app.tsx, but they're pure
helpers consumed by reactions/handlers/views. Move them under util/
so the src/ root is left to composition-root files.
