+++ packages/apps/navigator/src/tests/route-bounds.test.ts
@@ -1,4 +1,4 @@
-import { bboxesToCornerPairs, bboxToCornerPair } from '../route-bounds';
+import { bboxesToCornerPairs, bboxToCornerPair } from '../util/route-bounds';
 
 describe('bboxToCornerPair', () => {
   it.each([
