+++ packages/apps/navigator/src/tests/route-display.test.ts
@@ -3,7 +3,10 @@ import {
   defaultImperialOptions,
   defaultMetricOptions,
 } from '../components/text';
-import { sortedRoutePreviewIndices, toRouteSummary } from '../route-display';
+import {
+  sortedRoutePreviewIndices,
+  toRouteSummary,
+} from '../util/route-display';
 
 const stubRoute = (id: string): RouteWithSummary =>
   ({ id }) as RouteWithSummary;
