+++ packages/apps/navigator/src/util/route-bounds.ts
