+++ packages/apps/navigator/src/util/route-features.ts
